requirements.txt
pandas
numpy
regex
matplotlib
unidecode
tiktoken==0.7.0
pyyaml